langchain/chainlit_falcon_with_chat_ui/requirements.txt
chainlit==0.7.602
langchain==0.0.335
python-dotenv==1.0.0
huggingface_hub
nbconvert
